requirements.txt
amqp==5.1.1
arrow==1.3.0
asgiref==3.7.2
beautifulsoup4==4.12.2
billiard==4.1.0
binaryornot==0.4.4
black==23.9.1
celery==5.3.4
certifi==2023.7.22
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cookiecutter==2.5.0
cryptography==41.0.5
defusedxml==0.7.1
Django==4.2.6
django-allauth==0.58.2
django-browser-reload==1.12.1
django-debug-toolbar==4.2.0
django-filter==23.3
django-tailwind==3.6.0
djangorestframework==3.14.0
djangorestframework-api-key==3.0.0
googleapis-common-protos==1.61.0
grpcio==1.59.2
grpcio-status==1.59.2
gunicorn==21.2.0
idna==3.4
Jinja2==3.1.2
kombu==5.3.2
Markdown==3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==23.2
pathspec==0.11.2
pip-autoremove==0.10.0
platformdirs==3.11.0
prompt-toolkit==3.0.39
protobuf==4.25.0
psycopg2-binary==2.9.9
pycparser==2.21
Pygments==2.17.2
PyJWT==2.8.0
python-dateutil==2.8.2
python-slugify==8.0.1
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.0
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
text-unidecode==1.3
types-python-dateutil==2.8.19.14
tzdata==2023.3
urllib3==2.0.6
vine==5.0.0
wcwidth==0.2.8
